Revert "renamed AST parse file" for easier review of changes
This reverts commit a2524b781ca61c80a9ac49361a0fd411f7d9285b.
