+++ src/analysis/ASTnodefunctions.ts
